Feat: #108 DataStoreRepository 리팩토링
- DataStoreRepository 인터페이스 생성
- DataStoreRepositoryImpl 구현체 분리
- RepositoryModule에 의존성 주입 설정 업데이트
